ocr-integration-project/requirements.txt
flask
opencv-python
pytesseract
numpy
scikit-learn
nltk
Pillow
matplotlib
PyQt5
scikit-image

